// generated/src/app/api-models/pmb-memb-bill-hist.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbMembBillHist {
  customerType: string;
  customerId: string;
  invoiceNo: number;
  subscriberId: string;
  personNumber: string;
  recordType: string;
  detailFromDate: Date;
  detailThruDate: Date;
  planRiderCode: string;
  premiumAmt: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  familySize: number;
  spouseFlag: string;
  premiumStep: string;
}